DiscordBot/Community/requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0 